Refactor: GameController 관련 함수 logic에 분리
