+++ src/utils/MissionUtils.js
@@ -32,66 +32,10 @@ const getUserNumber = (computer, question) => {
   });
 };
 
-const countStrikeAndBall = (computers, user) => {
-  const users = useNumberToArray(user);
-
-  let countStrike = 0;
-  let countBall = 0;
-
-  users.forEach((number, index) => {
-    countStrike += isStrike(number, index, computers);
-    countBall += isBall(number, index, computers);
-  });
-  console.log(countStrike, countBall);
-  return { countStrike: countStrike, countBall: countBall };
-};
-
-const isStrike = (number, index, computers) => {
-  if (number === computers[index]) {
-    return 1;
-  }
-  return 0;
-};
-const isBall = (number, index, computers) => {
-  if (number !== computers[index] && computers.includes(number)) {
-    return 1;
-  }
-  return 0;
-};
-
-const isUserWin = (countStrike) => {
-  let userWin = countStrike === 3;
-  return userWin;
-};
-
 const printMessage = (message) => {
   MissionUtils.Console.print(message);
 };
 
-const printGameResultMessage = (countStrike, countBall) => {
-  let isUserWin = countStrike === MAX_CORRECT;
-  let message = makeMessage(countStrike, countBall);
-  printMessage(message);
-};
-
-const makeMessage = (countStrike, countBall) => {
-  if (countStrike === 3) {
-    return countStrike + STRIKE;
-  }
-
-  let message = "";
-  if (countBall > MIN_CORRECT) {
-    message += countBall + BALL + " ";
-  }
-  if (countStrike > MIN_CORRECT) {
-    message += countStrike + STRIKE;
-  }
-  if (countStrike === MIN_CORRECT && countBall === MIN_CORRECT) {
-    message = NOTTHING;
-  }
-
-  return message.trim();
-};
 const playAgain = () => {
   const computerNumbers = createComputerNumber(START_NUMBER, END_NUMBER);
   getUserNumber(computerNumbers, REQUIRE_NUMBER);
@@ -112,8 +56,5 @@ const retryOrEnd = () => {
 module.exports = {
   getUserNumber,
   setUserInput,
-  countStrikeAndBall,
-  isUserWin,
-  printGameResultMessage,
   printMessage,
 };
